Add unique post id to front matter.
The new post id is used for the disqus_identifier and is an integer.
Currently, I manually assign these in the front matter. post.authority
is now set on all post objects programmatically to be CONF.app.base_url,
e.g. "http://davidadrian.org". This way we don't have to pass extra
stuff to the template.